refactor:  self review changes
