+++ frontend/src/container/SideNav/menuItems.tsx
@@ -11,7 +11,6 @@ import {
 	Building2,
 	ChartArea,
 	Cloudy,
-	Coins,
 	DraftingCompass,
 	FileKey2,
 	Github,
@@ -365,13 +364,6 @@ export const settingsNavSections: SettingsNavSection[] = [
 				icon: <Sparkles size={16} />,
 				isEnabled: false,
 				itemKey: 'mcp-server',
-			},
-			{
-				key: ROUTES.LLM_OBSERVABILITY_MODEL_PRICING,
-				label: 'Model Pricing',
-				icon: <Coins size={16} />,
-				isEnabled: true,
-				itemKey: 'llm-observability-model-pricing',
 			},
 		],
 	},
